requirements.txt
attrs==20.3.0
click==7.1.2
fastjsonschema==2.14.5
Flask==1.1.2
importlib-metadata==3.3.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
packaging==20.8
pandas==1.2.0
pluggy==0.13.1
py==1.10.0
pymongo==3.11.2
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.0
